requirements.txt
Django==4.2.1
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
Pillow==9.5.0
pytz==2023.3
sqlparse==0.4.4
celery==5.3.1
redis== 4.6.0
django-celery-beat==2.5.0
drf-spectacular==0.26.3
